Modul19/mysite/requirements.txt
asgiref==3.7.1
attrs==23.1.0
certifi==2023.7.22
Django==4.2.1
django-debug-toolbar==4.1.0
django-filter==23.2
djangorestframework==3.14.0
drf-spectacular==0.26.4
inflection==0.5.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
Pillow==10.0.0
pytz==2023.3
PyYAML==6.0.1
referencing==0.30.0
rpds-py==0.9.2
sentry-sdk==1.28.1
sqlparse==0.4.4
typing_extensions==4.6.1
uritemplate==4.1.1
urllib3==2.0.4
